+++ de/factorization.py
@@ -90,10 +90,8 @@ def cellLineFactorization(cellLine):
     return w, eta, annotation[0].tolist()
 
 
-def cellLineComparison(cellLine1, cellLine2):
+def cellLineComparison(annotation1, annotation2):
     """Uses annotation list to generate an array of common genes between two cell lines"""
-    _, annotation1 = importLINCS(cellLine1)
-    _, annotation2 = importLINCS(cellLine2)
 
     annotation1 = annotation1[0].tolist()
     annotation2 = annotation2[0].tolist()
@@ -112,9 +110,11 @@ def cellLineComparison(cellLine1, cellLine2):
 def MatrixSubtraction(cellLine1, cellLine2):
     """Subtracts the w-matrices of two different cell lines and subtracts them.
     Then, it calculates the norm of the original matrices as well as difference matrix"""
+    _, annotation1 = importLINCS(cellLine1)
+    _, annotation2 = importLINCS(cellLine2)
     w1, _, _ = cellLineFactorization(cellLine1)
     w2, _, _ = cellLineFactorization(cellLine2)
-    index_list1, index_list2 = cellLineComparison(cellLine1, cellLine2)
+    index_list1, index_list2 = cellLineComparison(annotation1, annotation2)
 
     w1, _, _ = cellLineFactorization(cellLine1)
     w2, _, _ = cellLineFactorization(cellLine2)
@@ -137,7 +137,26 @@ def MatrixSubtraction(cellLine1, cellLine2):
     return norm1, norm2, diff_norm, w1_final, w2_final
 
 
-def cross_val(X, n=20):
+def mergedFitting(cellLine1, cellLine2):
+    """Given two cell lines, compute the cost of fitting each of them individually and the cost of fitting a shared w matrix."""
+    index_list1, index_list2 = cellLineComparison(cellLine1, cellLine2)
+
+    data1, _ = importLINCS(cellLine1)
+    data2, _ = importLINCS(cellLine2)
+    data1_df = pd.DataFrame(data1)
+    data2_df = pd.DataFrame(data2)
+    data1_edited = data1_df.iloc[index_list1, index_list1]
+    data2_edited = data2_df.iloc[index_list2, index_list2]
+    data1_final = data1_edited.values
+    data2_final = data2_edited.values
+    shared_data = [data1_final, data2_final]
+
+    w_shared, eta_list = factorizeEstimate(shared_data)
+
+    return w_shared, eta_list
+
+
+def split_data(X, n=20):
     """ Prepare the test and train data. """
     row = np.random.choice(X.shape[0], n, replace=False)
     col = np.random.choice(X.shape[1], n, replace=False)
